routelib: first attempt at zone-aware routes
The way the pool objects are being handled isn't up to my standards,
but there doesn't seem to be anything wrong with it.

Pools can be organized into "poolsets{}", which can be queried within a
route handler's setup code. 'local_zone' helps the route organize the
children before querying.

With this setup it's possible to mix zoned and non-zoned routes in the
same configuration, as well as override 'local_zone' on a per-route
basis if needed.