rlib: be slightly smarter for set queries
Need to continue to revise this but the behavior could be surprising so
lets do this one change now:

"set_name_blah" should work, and if "set_name" = { blah = {} } exists,
it will opportunistically use that.

next step is to probably fetch the final "_name" from the string instead
of simply the next one.